pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-log-gen"
version = "0.1.0"
description = "Synthetic process event log generator from YAML configuration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["process-mining", "event-log", "synthetic-data", "pm4py", "xes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pm4py>=2.7.0",
    "jsonschema>=4.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
    "pandas-stubs>=2.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
event-log-gen = "event_log_gen.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/event-log-gen"
Documentation = "https://github.com/yourusername/event-log-gen/blob/main/docs/USER_GUIDE.md"
Repository = "https://github.com/yourusername/event-log-gen"
Issues = "https://github.com/yourusername/event-log-gen/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=event_log_gen --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
